+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Some task parameters support environment variables and tilde expansion
 - `get_env` function can now return the environment variables defined in the task or file, which
 takes precedence over the system environment variables.
+- Make `echo` a builtin command
 
 ## v1.1.0 - 2023-06-01